+++ server/src/server_essentials.c
@@ -6,7 +6,7 @@
 
 #include "control_utils.h"
 
-void run_node(int32_t node_label) {
+void run_node(uint8_t node_label) {
 	char node_label_str[16];
 	int32_t len;
